Add a Zoom link to the episode details view
As presenters, we must open the episode details page before each
episode to put all our notes in the system.

After we're done, we may wish to jump straight into our recurring Zoom
meeting, which we use to record the episode.

To save us from having to find the link each time, I put it directly
in the episode details view for convenience.

The link is public anyway (listeners use it to call-in), so  this
presents no security risk.