Create an idiomatic Go directory structure
Move Go files responsible for the tyfloprzeglad web app  to
cmd/tyfloprzeglad.

This is in preparation for another related tool that will soon end up in this repo.

The only file remaining in the main package is data.go, which is
responsible for the data layer, and will also be used by the new tool.