Assign unique IDs to stories for editing and deleting
Stories now receive IDs (unique in their segment) to let us implement
story editing and deletion (which will be done soon).

Previously, we could only identify stories by position. Because stories
are often inserted in the middle of a segment, those positions might
change while the edit is in progress, causing us to edit the wrong story.

IDs are generated on story creation and never change.

When loading a data file with no IDs, they will be generated for all
existing stories in that file.